added rating logic; reworked books' fetching to include rating

+++ src/main/java/com/neoxenus/webnovelreader/book/dto/projection/BookRatingAverages.java
@@ -1,7 +1,7 @@
 package com.neoxenus.webnovelreader.book.dto.projection;
 
 public interface BookRatingAverages {
-    Integer getRatingCount();
+    Long getRatingCount();
     Double getAvgStoryDevelopment();
     Double getAvgWritingQuality();
     Double getAvgWorldBackground();
